fix: bring back new button
